! #14 ClusterRaftClients now require the role 'raft' (configurable) to run From now on, raft cluster actors will only run on nodes that have a specific role. Should be useful in larger deployments, when we only want raft to run on 5 (and not all 200) nodes in the cluster.
Decided against making this role check optional
(configurable), because it seems pretty core to designing a cluster,
and one can always add the role to all nodes during deployment.

@patriknw Do you think the `system.stop(self)` is too "heavy" or a good
idea? I personaly like it, but it's not really required - we can simply
not talk to nodes without the 'raft' role (ClusterRaftActor:50). But
without the stop call, there would be "raft zombies" running, so I
went ahead with this impl for now. Thanks for the idea btw!