! split ClusterRaftActor and RaftActor into 2 real actors Mostly thanks to Patriks review comments.
The cluster actor now is a "proxy", which is able to take care of all
the cluster related management - such as discovering nodes, and trying
to select members on other nodes etc. Not sure about deathwatch (if the
cluster actor should kill itself or not (currently it does) when the
raft one dies.).

Initial steps towards auto discovery - what is missing is supporting
these messages "during" the lifecycle, not only in `Init`, which will be
task #18.

Also moved towards #18 and #17, by making things more configurable.

Roles seem to be covered for now; resolves #14