+ #15 RaftClientActor is able to find the Leader, and stashes until a leader is found; when it is found, it will send messages to it.
TODO: tests for that a leader has died, and a new one is now being elected

Makes #19 painfuly visible, having 2 actor refs passed around makes
things not really nice, because of the matchIndex etc being actor ref
based. And it can happen in a cluster, that the "raft" ref gets sent
around, or the "cluster" ref. TODO: Decide either, on a way to always
use the right ref; or kill the separate actors idea and make cluster
awareness a trait again, like it was previously.